+++ src/main/java/ru/vlad/altituda/controllers/SubCategoriesController.java
@@ -5,6 +5,7 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
+import ru.vlad.altituda.JDBC.Registration;
 import ru.vlad.altituda.Model.SubCategories;
 import ru.vlad.altituda.dao.SubCategoriesDAO;
 
@@ -23,41 +24,25 @@ public class SubCategoriesController {
 
     @GetMapping()
     public String index(Model model){
+        if (Registration.isRegistration() != true) return "redirect:/login";
         model.addAttribute("subCategories",subCategoriesDAO.index());
+        model.addAttribute("user",Registration.getUsers());
         return "subCategories/listSubCategories";
-    }
-
-
-    @GetMapping("/categories")
-    public String categories(){
-        return "/categories";
-    }
-    @GetMapping("/users")
-    public String users(){
-        return "/users";
-    }
-    @GetMapping("/order")
-    public String order(){
-        return "/order";
-    }
-    @GetMapping("/product")
-    public String product(){
-        return "/product";
-    }
-    @GetMapping("/producer")
-    public String producer(){
-        return "/producer";
     }
 
     @GetMapping("/{id}")
     public String show(@PathVariable("id") int id, Model model){
+        if (Registration.isRegistration() != true) return "redirect:/login";
         model.addAttribute("subCategories",subCategoriesDAO.show(id));
+        model.addAttribute("user",Registration.getUsers());
         return "redirect:/subCategories";
     }
 
 
     @GetMapping("/new")
-    public String newPerson(@ModelAttribute("subCategories") SubCategories subCategories){
+    public String newPerson(@ModelAttribute("subCategories") SubCategories subCategories,Model model){
+        if (Registration.isRegistration() != true) return "redirect:/login";
+        model.addAttribute("user",Registration.getUsers());
         return "subCategories/new";
     }
 
@@ -65,6 +50,7 @@ public class SubCategoriesController {
     @PostMapping()
     public String create(@ModelAttribute("subCategories") @Valid SubCategories subCategories,
                          BindingResult bindingResult){
+        if (Registration.isRegistration() != true) return "redirect:/login";
         if (bindingResult.hasErrors()){
             return "subCategories/new";
         }
@@ -74,13 +60,16 @@ public class SubCategoriesController {
 
     @GetMapping("/{id}/edit")
     public String edit(Model model, @PathVariable("id") int id){
+        if (Registration.isRegistration() != true) return "redirect:/login";
         model.addAttribute("subCategories",subCategoriesDAO.show(id));
+        model.addAttribute("user",Registration.getUsers());
         return "subCategories/edit";
     }
 
     @PostMapping ("/{id}")
     public String update(@ModelAttribute("subCategories") @Valid SubCategories subCategories,BindingResult bindingResult,
                          @PathVariable("id")int id){
+        if (Registration.isRegistration() != true) return "redirect:/login";
         if (bindingResult.hasErrors()){
             return "subCategories/edit";
         }
@@ -91,6 +80,7 @@ public class SubCategoriesController {
 
     @GetMapping("/{id}/delete")
     public String delete(@PathVariable("id") int id){
+        if (Registration.isRegistration() != true) return "redirect:/login";
         subCategoriesDAO.delete(id);
         return "redirect:/subCategories";
     }
