+++ src/main/java/ru/vlad/altituda/controllers/UsersController.java
@@ -5,7 +5,10 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
+import ru.vlad.altituda.JDBC.Registration;
 import ru.vlad.altituda.Model.Users;
+import ru.vlad.altituda.dao.ProductDAO;
+import ru.vlad.altituda.dao.SubCategoriesDAO;
 import ru.vlad.altituda.dao.UserDAO;
 
 import javax.validation.Valid;
@@ -16,48 +19,35 @@ public class UsersController {
 
     @Autowired
     private final UserDAO userDAO;
+    @Autowired
+    private final ProductDAO productDAO;
 
-    public UsersController(UserDAO userDAO) {
+    public UsersController(UserDAO userDAO, ProductDAO productDAO) {
         this.userDAO = userDAO;
+        this.productDAO = productDAO;
     }
 
     @GetMapping()
     public String index(Model model){
+        if (Registration.isRegistration() != true) return "redirect:/login";
+        model.addAttribute("user",Registration.getUsers());
         model.addAttribute("users", userDAO.index());
         return "users/listUsers";
     }
 
-
-    @GetMapping("/subCategories")
-    public String subCategories(){
-        return "/subCategories";
-    }
-    @GetMapping("/categories")
-    public String categories(){
-        return "/categories";
-    }
-    @GetMapping("/order")
-    public String order(){
-        return "/order";
-    }
-    @GetMapping("/product")
-    public String product(){
-        return "/product";
-    }
-    @GetMapping("/producer")
-    public String producer(){
-        return "/producer";
-    }
-
     @GetMapping("/{email}")
     public String show(@PathVariable("email") String email, Model model){
+        if (Registration.isRegistration() != true) return "redirect:/login";
+        model.addAttribute("user",Registration.getUsers());
         model.addAttribute("users", userDAO.show(email));
         return "redirect:/users";
     }
 
 
     @GetMapping("/new")
-    public String newPerson(@ModelAttribute("users") Users users){
+    public String newPerson(@ModelAttribute("users") Users users,Model model){
+        if (Registration.isRegistration() != true) return "redirect:/login";
+        model.addAttribute("user",Registration.getUsers());
         return "users/new";
     }
 
@@ -65,6 +55,7 @@ public class UsersController {
     @PostMapping()
     public String create(@ModelAttribute("users") @Valid Users users,
                          BindingResult bindingResult){
+        if (Registration.isRegistration() != true) return "redirect:/login";
         if (bindingResult.hasErrors()){
             return "users/new";
         }
@@ -74,6 +65,8 @@ public class UsersController {
 
     @GetMapping("/{email}/edit")
     public String edit(Model model, @PathVariable("email") String email){
+        if (Registration.isRegistration() != true) return "redirect:/login";
+        model.addAttribute("user",Registration.getUsers());
         model.addAttribute("users", userDAO.show(email));
         return "users/edit";
     }
@@ -81,6 +74,7 @@ public class UsersController {
     @PostMapping ("/{email}")
     public String update(@ModelAttribute("users") @Valid Users users,BindingResult bindingResult,
                          @PathVariable("email") String email){
+        if (Registration.isRegistration() != true) return "redirect:/login";
         if (bindingResult.hasErrors()){
             return "users/edit";
         }
@@ -88,9 +82,9 @@ public class UsersController {
         return "redirect:/users";
     }
 
-
     @GetMapping("/{email}/delete")
     public String delete(@PathVariable("email") String email){
+        if (Registration.isRegistration() != true) return "redirect:/login";
         userDAO.delete(email);
         return "redirect:/users";
     }
